requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
atomicwrites==1.3.0
attrs==19.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
enum34==1.1.10
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
more-itertools==8.2.0
namedlist==1.7
packaging==20.3
pathlib2==2.3.5
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
pytest-html==1.19.0
pytest-metadata==1.8.0
pytest-mock==1.9.0
python-dateutil==2.8.1
PyYAML==3.12
requests==2.23.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==1.2.0
